+++ app/(navigatable)/(goals)/goals/[id]/page.tsx
@@ -17,7 +17,7 @@ import {
 } from './../../../../../lib/data-model-types'
 
 const GoalIndex = ({ params: { id } }: { params: { id: string } }) => {
-  const { setActiveGoalId, activeWorkspaceDetails } = useAppContext()
+  const { setProjectIdForConfiguration, setActiveGoalId, activeWorkspaceDetails } = useAppContext()
 
   const [goal, setGoal] = useState<Goal | null>(null)
   const [health, setHealth] = useState<ProjectHealth | null>(null)
@@ -37,6 +37,13 @@ const GoalIndex = ({ params: { id } }: { params: { id: string } }) => {
       )
 
       setHealth(health)
+
+      const project = await getResource<Goal>(
+        `/goals/${id}/parentproject`,
+        `Failed to get a success response when fetching parent project for goal '${id}'`,
+      )
+      setProjectIdForConfiguration(project.id)
+
     })()
   }, [id])